create remove confirmation

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -6,14 +6,14 @@ import { MatDialog } from '@angular/material/dialog';
 import { ErrorDialogComponent } from '../../../shared/components/error-dialog/error-dialog.component';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
   templateUrl: './courses.component.html',
-  styleUrls: ['./courses.component.css']
+  styleUrls: ['./courses.component.css'],
 })
-export class CoursesComponent implements OnInit{
-
+export class CoursesComponent implements OnInit {
   courses$: Observable<Course[]> | null = null;
 
   constructor(
@@ -21,51 +21,56 @@ export class CoursesComponent implements OnInit{
     public dialog: MatDialog,
     private router: Router,
     private route: ActivatedRoute,
-    private _snackBar: MatSnackBar,
-  ){
+    private _snackBar: MatSnackBar
+  ) {
     this.refresh();
   }
 
-  refresh(){
-    this.courses$ = this.coursesService.list()
-    .pipe(
-      catchError(error => {
+  refresh() {
+    this.courses$ = this.coursesService.list().pipe(
+      catchError((error) => {
         this.onError('Erro ao carregar cursos.');
-        return of([])
+        return of([]);
       })
     );
   }
 
   onError(erroMsg: string) {
     this.dialog.open(ErrorDialogComponent, {
-      data: erroMsg
+      data: erroMsg,
     });
   }
 
-  ngOnInit(): void {
+  ngOnInit(): void {}
 
-  }
-
-  onAdd(){
-    this.router.navigate(['new'], {relativeTo: this.route});
+  onAdd() {
+    this.router.navigate(['new'], { relativeTo: this.route });
   }
 
   onEdit(course: Course): void {
-    this.router.navigate(['edit', course._id ], {relativeTo: this.route});
+    this.router.navigate(['edit', course._id], { relativeTo: this.route });
   }
 
-  onRemove(course: Course){
-    this.coursesService.remove(course._id).subscribe({
-      next: () => {
-        this.refresh();
-        this._snackBar.open('Course removed successfully.', 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
+  onRemove(course: Course) {
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Are you sure you want to remove this course?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe({
+          next: () => {
+            this.refresh();
+            this._snackBar.open('Course removed successfully.', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
+          },
+          error: () => {
+            this.onError('Error when trying to remove course.');
+          },
         });
-      },
-      error: () => {
-        this.onError("Error when trying to remove course.");
       }
     });
   }
